fix(ci): pay the source-size ratchet down instead of raising it
Three CI reds, all mine.

check_source_size_budget. cmd_doctor.ts sits ~2,100 lines past the 1,500-line
threshold, so every line added to it is new excess, and that ratchet turns one
way while its test asserts baseline == live in BOTH directions. The branch came
in at +137.

Paid down in two moves rather than raised. The strict semantics, threshold,
kill switch, payload helper and argv fragments moved to _cli/_doctor_strict.ts
(~160 lines, under the cap, costs nothing), taking cmd_doctor.ts to +15. Then
the 37-line drift renderer moved to _cli/_doctor_emit.ts. Net -21, and the
baseline is lowered 18,230 -> 18,209 because the gate and its test both require
the lowering commit to carry it.

The renderer was chosen deliberately: pure rendering, two dependencies, no
argparse involvement and therefore none of the byte-parity risk the parser
surface carries against the retired Python implementation. Nothing outside
cmd_doctor.ts imported it -- cmd_prune.ts has a same-named local of its own and
is untouched -- and the export surface is preserved.

An AI council on the 15-line residual SPLIT: one seat for raising the baseline
with a payback obligation, one for descoping the feature until the file is
reduced. Neither was taken. A parallel branch demonstrated the same day that
paying down is achievable here, which is the evidence the split lacked, and
this gate's own test comment already prescribes it: a commit that splits a
god-file lowers the excess and must carry the lowered baseline.

lint_code_comments rejected the gate header's incident narration as a
provenance comment. The evidence belongs in the roadmap, which has it; the
reach limit that step 1.3 mandates stays, rewritten as a constraint.

lint_canonical_terms was 5 over on `artefact` in the agents-layout edit. Fixed
line-scoped -- three further instances in that file predate this change and are
left alone.

Behaviour re-verified after the move: plain exit 0, --strict 1, kill switch 0
with the suppression announced, unknown level 2, can_proceed false in --json.
135 tests across the three doctor suites.
